Fix for race condition
We ran into a race condition where a viewController can subscribe upon loading, update and before the subscription gets added to the subscription list, an event can be fired with an update. The result is that the ViewController doesn't get updated for that event.
In our case, we are blocking the UI while we fire a command to tell us what segue to fire and that command is finishing before the viewController is added to the list of subscriptions in certain conditions, leaving the ViewController indefinitely frozen. From our testing, this fix should rectify that problem.